Remove another warning message.
git-svn-id: file://localhost/home/tolik/svnrepo/trunk@954 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ opencv/src/highgui/window_carbon.cpp
@@ -789,8 +789,8 @@ static pascal OSStatus windowEventHandler(EventHandlerCallRef nextHandler, Event
                 Rect structure, content;
                 GetWindowBounds(theWindow, kWindowStructureRgn, &structure);
                 GetWindowBounds(theWindow, kWindowContentRgn, &content);
-                lx = point.x - content.left + structure.left;
-                ly = point.y - window->trackbarheight  - content.top + structure.top; /* minus la taille des trackbars */
+                lx = (int)point.x - content.left + structure.left;
+                ly = (int)point.y - window->trackbarheight  - content.top + structure.top; /* minus la taille des trackbars */
                 if (window->flags & CV_WINDOW_AUTOSIZE) {//FD
                                                          //printf("was %d,%d\n", lx, ly);
                     /* scale the mouse coordinates */
